Moved parts of tertiary out into new components
